requirements.txt
fastapi
pydantic
uvicorn[standard]
pytest
sqlalchemy
pymysql
PyYAML
cryptography
httpx